greenhouse_resume_content_sync/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
psycopg2-binary>=2.9.0
